# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin Panel",
        "description": "With Passport login and datatable custom filter",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/laravel-example-admin/public/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/reset-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Reset Password Using OTP",
                "operationId": "reset-password",
                "requestBody": {
                    "description": "Reset Password",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "otp",
                                    "password",
                                    "confirm_password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Enter email",
                                        "type": "string"
                                    },
                                    "otp": {
                                        "description": "Enter otp",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Enter password",
                                        "type": "string"
                                    },
                                    "confirm_password": {
                                        "description": "Enter confirm password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}